chore: move stats api to app route
